+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/android/android_build_prechecks.rb
@@ -10,25 +10,20 @@ module Fastlane
         # Verify that the current branch is a release branch. Notice that `ensure_git_branch` expects a RegEx parameter
         ensure_git_branch(branch: '^release/') unless other_action.is_ci
 
-        has_alpha_version = params[:has_alpha_version]
-        project_root_folder = params[:project_root_folder]
-        project_name = params[:project_name]
-        build_gradle_path = params[:build_gradle_path] || (File.join(project_root_folder || '.', project_name, 'build.gradle') unless project_name.nil?)
-        version_properties_path = params[:version_properties_path] || File.join(project_root_folder || '.', 'version.properties')
+        build_gradle_path = params[:build_gradle_path]
+        version_properties_path = params[:version_properties_path]
 
         message = ''
         unless !params[:beta] && !params[:final]
           beta_version = Fastlane::Helper::Android::VersionHelper.get_release_version(
             build_gradle_path: build_gradle_path,
-            version_properties_path: version_properties_path,
-            has_alpha_version: has_alpha_version
+            version_properties_path: version_properties_path
           )
         end
         if params[:alpha]
           alpha_version = Fastlane::Helper::Android::VersionHelper.get_alpha_version(
             build_gradle_path: build_gradle_path,
-            version_properties_path: version_properties_path,
-            has_alpha_version: has_alpha_version
+            version_properties_path: version_properties_path
           )
         end
 
@@ -86,19 +81,12 @@ module Fastlane
                                        description: 'Path to the build.gradle file',
                                        type: String,
                                        optional: true,
-                                       conflicting_options: %i[project_name
-                                                               project_root_folder
-                                                               version_properties_path]),
+                                       conflicting_options: [:version_properties_path]),
           FastlaneCore::ConfigItem.new(key: :version_properties_path,
                                        description: 'Path to the version.properties file',
                                        type: String,
                                        optional: true,
-                                       conflicting_options: %i[build_gradle_path
-                                                               project_name
-                                                               project_root_folder]),
-          Fastlane::Helper::Deprecated.project_root_folder_config_item,
-          Fastlane::Helper::Deprecated.project_name_config_item,
-          Fastlane::Helper::Deprecated.has_alpha_version_config_item,
+                                       conflicting_options: [:build_gradle_path]),
         ]
       end